backend/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
asyncio==4.0.0
blinker==1.9.0
certifi==2025.10.5
click==8.3.0
colorama==0.4.6
Flask==3.1.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
letta-client==0.1.324
MarkupSafe==3.0.3
pydantic==2.12.3
pydantic_core==2.41.4
python-dotenv==1.1.1
sniffio==1.3.1
typing-inspection==0.4.2
typing_extensions==4.15.0
Werkzeug==3.1.3
